pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
]
name = "nomad-plugin-spectroscopy"
dynamic = ["version"]
description = "NOMAD plugin for parsing and managing spectroscopy data (IR, Raman, UV-Vis)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Marzieh Saeedimasine", email = "marzieh.saeedimasine@gmail.com" },
]
maintainers = [
    { name = "Marzieh Saeedimasine", email = "marzieh.saeedimasine@gmail.com" },
]
license = { text = "MIT" }
dependencies = [
    "nomad-lab>=1.4.0",
    "pydantic",
]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-plugin-spectroscopy"
Documentation = "https://fairmat-nfdi.github.io/nomad-plugin-spectroscopy"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-plugin-spectroscopy/issues"
Changelog = "https://github.com/FAIRmat-NFDI/nomad-plugin-spectroscopy/releases"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "structlog>=23.1.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "pymdown-extensions>=10.0.0",
    "mkdocs-click>=0.8.1",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "pymdown-extensions>=10.0.0",
    "mkdocs-click>=0.8.1",
]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "C4",  # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/nomad_plugin_spectroscopy --cov-report=html --cov-report=term"
asyncio_mode = "auto"
